(PC-17812) favorites: put api hooks in /api folder
	+ export types in type.ts file
