+++ src/features/favorites/api/index.ts
@@ -1,0 +1,5 @@
+export { useFavoritesCount } from './useFavoritesCount'
+export { useFavorites } from './useFavorites'
+export { useFavorite } from './useFavorite'
+export { useAddFavorite } from './useAddFavorite'
+export { useRemoveFavorite } from './useRemoveFavorite'
