+++ src/features/favorites/api/types.ts
@@ -1,0 +1,5 @@
+import { FavoriteResponse } from 'api/gen'
+
+export interface FavoriteMutationContext {
+  previousFavorites: Array<FavoriteResponse>
+}
