+++ src/features/favorites/components/FavoritesResults.tsx
@@ -4,18 +4,18 @@ import styled from 'styled-components/native'
 
 import { FavoriteOfferResponse, FavoriteResponse } from 'api/gen'
 import { BookingOfferModal } from 'features/bookOffer/pages/BookingOfferModal'
+import { useFavorites } from 'features/favorites/api'
 import { Sort } from 'features/favorites/components/Buttons/Sort'
 import { Favorite } from 'features/favorites/components/Favorite'
 import { NoFavoritesResult } from 'features/favorites/components/NoFavoritesResult'
 import { NumberOfResults } from 'features/favorites/components/NumberOfResults'
-import { FavoriteSortBy } from 'features/favorites/pages/FavoritesSorts'
 import { useFavoritesState } from 'features/favorites/pages/FavoritesWrapper'
 import {
   sortByAscendingPrice,
   sortByDistanceAroundMe,
   sortByIdDesc,
 } from 'features/favorites/pages/helpers/sorts'
-import { useFavorites } from 'features/favorites/pages/useFavorites'
+import { FavoriteSortBy } from 'features/favorites/types'
 import { useAvailableCredit } from 'features/home/services/useAvailableCredit'
 import { useUserProfileInfo } from 'features/profile/api'
 import { useGeolocation, GeoCoordinates } from 'libs/geolocation'
