+++ src/features/favorites/types.ts
@@ -1,0 +1,1 @@
+export type FavoriteSortBy = 'RECENTLY_ADDED' | 'ASCENDING_PRICE' | 'AROUND_ME'
